refactor: move attack state to combat module
